User Experience Fixes
Made the screen clear when a user makes a selection from the main menu.
Fixed the generation of matrices so they now generate as identity or the equivalent.
Made the print matrix function exit automatically if there are no matrices to print.
Fixed getting stuck if username or password were not known